MQTTMonitor/requirements.txt
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
freeze==1.0.10
idna==2.8
influxdb==5.2.3
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.3
paho-mqtt==1.4.0
pbr==5.4.3
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
scipy==1.3.1
six==1.12.0
stevedore==1.31.0
urllib3==1.25.6

